Side-by-side images for students via simple CLI
This commit is as far as I made it in 2025. It overall went quite well
and a bunch of students said it was their favorite project of the year.
The problem was that 2-3 of the weaker students fell behind. This
builds on the experience I had last year where a student got VERY far
behind and required a lot of work on my part to bring them back on
track.

What is TBD is the following:
- Updating the worksheets to have forced screenshots so I can better
  monitor this
- Updating the scripts so it's initially "by sight" but then later by
  computation telling students what their grade is